[MoE] Fix GptOss expert output dtype for non-bf16 models
`GptOssGroupedExperts._experts_forward` returned the `torch._grouped_mm` output (bf16) without casting back to the input dtype, unlike the base `GroupedExperts._experts_forward`, which ends with `.type_as(x_RD)`. `LocalTokenDispatcher.combine` allocates the output buffer with `torch.zeros_like(x_TD)` (the model dtype) and `dispatch` preserves the input dtype, so at a non-bf16 dtype (e.g. `float32`) the bf16 expert output mismatched the fp32 buffer and `deterministic_scatter_add` raised `scatter(): Expected self.dtype to be equal to src.dtype`. This is why the GptOss GraphTrainer trace tests (which run at `float32`) were skipped.

Fix: cast GptOss's expert output with `.type_as(x_RD)`, mirroring `GroupedExperts`. It's a no-op in bf16 (no change to existing numerics or bitwise-determinism hashes) and resolves the mismatch otherwise. Re-enables `TestTraceModels.test_gpt_oss` and `TestTraceFSDP.test_gpt_oss_fsdp`.

Testing:
- Confirmed the dtype flow: the base experts cast via `.type_as(x_RD)` and dispatch preserves dtype, so only GptOss (missing the cast) returned bf16. With `.type_as(x_RD)`, an fp32 run yields fp32 expert output (matching the combine buffer) and bf16 is a no-op (returns the same tensor).
- Full `gpt_oss` trace tests (`pytest torchtitan/experiments/graph_trainer/tests/test_trace_module.py -k gpt_oss`) plus the bitwise-determinism guardrail are covered by CI.

+++ torchtitan/models/gpt_oss/moe.py
@@ -127,6 +127,11 @@ class GptOssGroupedExperts(Module):
             [num_tokens_per_expert_E, tail_slack]
         ).long()
 
+        # torch._grouped_mm needs bf16 inputs, so both matmuls run in bf16 and the
+        # bias-adds/swiglu stay in bf16 too (matching GroupedExperts; only the bias is
+        # new here). The output is cast back to the input dtype at the end via
+        # .type_as(x_RD). Keeping intermediates in bf16 is fine since real usage is bf16.
+
         # G = gate+up dimension (2*F)
         h_RG = torch._grouped_mm(
             x_RD.bfloat16(),
@@ -155,7 +160,8 @@ class GptOssGroupedExperts(Module):
             num_tokens_per_expert_long, dim=0, output_size=x_RD.shape[0]
         )
         b2_RD = ScaleBiasForward.apply(b2_RD, tp_degree)
-        return h_RD + b2_RD.to(h_RD.dtype)
+        # Cast bf16 result back to the input dtype (see note above), like GroupedExperts.
+        return (h_RD + b2_RD.to(h_RD.dtype)).type_as(x_RD)
 
     def forward(
         self,
